requirements.txt
# Core ML Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0

# Deep Learning
torch>=1.9.0
transformers>=4.9.0

# Text Processing
nltk>=3.6.0
spacy>=3.1.0

# Utilities
tqdm>=4.62.0
pyyaml>=5.4.0
joblib>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Notebook Support
jupyter>=1.0.0
notebook>=6.4.0